Moves Category into a separate class
